Handle renegotiations for subscribers with the HPB
In some cases the HPB will now send an updated offer to renegotiate an
already existing subscriber connection. Therefore, if the offer received
from the signaling server contains "update: true" it needs to be handled
by the current RTCPeerConnection, rather than being seen as stale and
ending it before creating a new one.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>